pyproject.toml
[project]
name = "dmserv"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Zowie", email = "git@zowie.cx" }
]
dependencies = [
    "py-cord[speed]>=2.5.0",
    "python-dotenv>=1.0.1",
    "SQLAlchemy>=2.0.28",
    "pluralkit>=1.1.6",
]
readme = "README.md"
requires-python = ">= 3.8"

[project.scripts]
dmserv = "dmserv.cmd.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "black>=24.3.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/dmserv"]
